+++ Helper/Config.php
@@ -4,7 +4,6 @@ namespace Dintero\Checkout\Helper;
 
 use Dintero\Checkout\Model\Api\Client;
 use Dintero\Checkout\Model\Dintero;
-use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Helper\Context;
 use Magento\Framework\Encryption\EncryptorInterface;
@@ -86,6 +85,16 @@ class Config extends AbstractHelper
     const XPATH_IS_EMBEDDED = 'payment/dintero/is_embedded';
 
     /*
+     * Embed Type
+     */
+    const XPATH_EMBED_TYPE = 'payment/dintero/embed_type';
+
+    /*
+     * Embed Type
+     */
+    const XPATH_IS_POPOUT = 'payment/dintero/is_popout';
+
+    /*
      * XPATH Express checkout enabled
      */
     const XPATH_IS_EXPRESS = 'payment/dintero/is_express';
@@ -157,10 +166,11 @@ class Config extends AbstractHelper
      * @param EncryptorInterface $encryptor
      */
     public function __construct(
-        Context $context,
-        EncryptorInterface $encryptor,
+        Context               $context,
+        EncryptorInterface    $encryptor,
         StoreManagerInterface $storeManager
-    ) {
+    )
+    {
         parent::__construct($context);
         $this->encryptor = $encryptor;
         $this->storeManager = $storeManager;
@@ -574,7 +584,7 @@ class Config extends AbstractHelper
      */
     public function getSessionExpirationDays()
     {
-        return  $this->scopeConfig->getValue(self::XPATH_SESSION_EXP_DAY);
+        return $this->scopeConfig->getValue(self::XPATH_SESSION_EXP_DAY);
     }
 
     /**
@@ -622,5 +632,39 @@ class Config extends AbstractHelper
             ',',
             $this->scopeConfig->getValue(self::XPATH_UNSPECIFIED_METHODS, ScopeInterface::SCOPE_STORE, $scopeCode)
         );
+    }
+
+    /**
+     * Retrieve embed type
+     *
+     * @param string $scopeCode
+     * @return string
+     */
+    public function getEmbedType($scopeCode = null)
+    {
+        $embedType = $this->scopeConfig->getValue(self::XPATH_EMBED_TYPE, ScopeInterface::SCOPE_STORE, $scopeCode);
+        return $embedType === Client::TYPE_EXPRESS ? Client::TYPE_EXPRESS : Client::TYPE_EMBEDDED;
+    }
+
+    /**
+     * Check if embedded express is enabled
+     *
+     * @param string $scopeCode
+     * @return bool
+     */
+    public function isEmbeddedExpress($scopeCode = null)
+    {
+        return $this->getEmbedType($scopeCode) === Client::TYPE_EXPRESS;
+    }
+
+    /**
+     * Check if embedded checkout popout mode is enabled
+     *
+     * @param $scopeCode
+     * @return bool
+     */
+    public function isPopOut($scopeCode = null)
+    {
+        return $this->scopeConfig->isSetFlag(self::XPATH_IS_POPOUT, ScopeInterface::SCOPE_STORE, $scopeCode);
     }
 }