+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+v1.8.1
+- added embedded popout functionality
+- added embedded express functionality
+- added embedded express with popout functionality
+- implemented coupon codes support for express checkout
+
 v1.7.15
 - Fixed item calculation amount